ml_architecture/requirements_ml.txt
# Core ML Framework
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1

# Transformers and NLP
transformers==4.35.2
tokenizers==0.15.0
spacy==3.8.7
nltk==3.8.1

# Traditional ML
scikit-learn==1.3.2
scipy==1.15.3
numpy==1.26.2
pandas==2.1.3

# Model Management
mlflow==2.8.1
dvc==3.30.1
optuna==3.4.0

# Data Processing
datasets==2.15.0
accelerate==0.24.1
sentencepiece==0.1.99

# Evaluation and Metrics
evaluate==0.4.1
rouge-score==0.1.2
bleu==0.1.0

# Utilities
tqdm==4.67.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# API and Web
fastapi==0.115.13
uvicorn==0.32.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Monitoring and Logging
wandb==0.16.1
tensorboard==2.15.1
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Production
gunicorn==21.2.0
docker==6.1.3
kubernetes==28.1.0

# Additional NLP
textstat==0.7.3
language-tool-python==2.7.1
pyenchant==3.2.2

# Feature Engineering
feature-engine==1.6.2
category-encoders==2.6.3

# Model Optimization
onnx==1.15.0
onnxruntime==1.16.3
torchserve==0.8.2

# Data Validation
great-expectations==0.17.23
pandera==0.17.2

# Configuration
hydra-core==1.3.2
omegaconf==2.3.0
pydantic==2.11.7

# Async and Performance
aiohttp==3.9.1
asyncio-mqtt==0.16.1
celery==5.3.4

# Security
cryptography==41.0.8
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.10
pdf2image==1.17.0
PyMuPDF==1.24.1 